% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalityOS.R
\name{get_scientific_object_modalities}
\alias{get_scientific_object_modalities}
\title{Scientific Object Modalities}
\usage{
get_scientific_object_modalities(
  host,
  user,
  password,
  experiment_uri,
  scientific_object_type,
  header_type = "both"
)
}
\arguments{
\item{host}{url of the hosting opensilex service}

\item{user}{opensilex user}

\item{password}{opesnielx user's password}

\item{experiment_uri}{uri of the targeted experiment}

\item{scientific_object_type}{uri of the targeted scientific object type}

\item{header_type}{header type: one of "initial","translated" or "both}
}
\description{
Scientific Object Modalities
}
