% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{get_token}
\alias{get_token}
\title{Get Token}
\usage{
get_token(host, user = "guest@opensilex.org", password = "guest")
}
\arguments{
\item{host}{url of the hosting opensilex service}

\item{user}{opensilex user}

\item{password}{opensilex user's password}
}
\value{
A string token, used by every subsequent call
}
\description{
Function use to get the token for a user
}
