% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{GetToken}
\usage{
getToken(host, user = "admin@opensilex.org", password = "admin")
}
\arguments{
\item{host}{url of the hosting opensilex service}

\item{user}{opensilex user}

\item{password}{opesnielx user's password}
}
\value{
A string token, used by every subsequent call
}
\description{
GetToken
}
