% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalityOS.R
\name{os_modality}
\alias{os_modality}
\title{Scientific Object Modalities}
\usage{
os_modality(host, user, password, experiment_uri, scientific_object_type)
}
\arguments{
\item{host}{url of the hosting opensilex service}

\item{user}{opensilex user}

\item{password}{opesnielx user's password}

\item{experiment_uri}{uri of the targeted experiment}

\item{scientific_object_type}{uri of the targeted scientific object type}
}
\description{
Scientific Object Modalities
}
