% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HttpStatusHelper.R
\name{parse_status}
\alias{parse_status}
\title{parse_status}
\usage{
parse_status(response_object)
}
\arguments{
\item{response_object}{object of class httr:response}
}
\value{
The same response_object passed in
}
\description{
Parse an http response and display a message and a log
}
